feat: 사각형 안에 들어가는지 검사하는 메소드 및 API를 BigDecimal 타입을 사용하여 계산하계 변경

+++ src/main/java/com/team5/capstone/mju/apiserver/web/controller/ParkingLotController.java
@@ -15,6 +15,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Slf4j
@@ -60,10 +61,10 @@ public class ParkingLotController {
     )
     @GetMapping("/parking-lots/rectangle")
     public ResponseEntity<List<ParkingLotResponseDto>> getParkingLotsInDisplay(
-            @RequestParam double x1, // 왼쪽 위 위도
-            @RequestParam double y1, // 왼쪽 위 경도
-            @RequestParam double x2, // 오른쪽 아래 위도
-            @RequestParam double y2 // 오른쪽 아래 경도
+            @RequestParam BigDecimal x1, // 왼쪽 위 위도
+            @RequestParam BigDecimal y1, // 왼쪽 위 경도
+            @RequestParam BigDecimal x2, // 오른쪽 아래 위도
+            @RequestParam BigDecimal y2 // 오른쪽 아래 경도
     ) {
         return ResponseEntity.ok(null);
     }
